Implement plugin functionality in Go
A proof of concept for implementing a Buildkite plugin in a different
language. This came out of realising that implementing a report via jq
wasn't that maintainable, and other options just highlight that Bash
isn't a good fit as the needs became more complicated.

The idea here is to ship the plugin as a binary via Docker, skipping the
need to build on the agent.

Don't fail a build when the plugin fails