fix: default summary state to have empty counts
This reduces the size of the rendered output when there are no vulnerabilities, as it will
no longer render 0 counts by default.

This behaviour was introduced recently with the ignore functionality, and results in a lot of
real estate at the top of a build being taken up for no purpose.