ci: funcorder's value is debatable
The linter funcorder suggests that constructors for structs should be
placed before struct methods. This seems like a good idea, though I am
not privy to what problem this is exactly trying to solve in this
repository. Currently, this is being maintained by Culture Amp's SRE
team, and my perception is that we don't have a strong enough opinion
on this that the linter is necessary.