feat: exit with a specific exit code if the scan times out (#128)
Occasionally the scan results will take longer than the timeout
period to complete. This is not necessarily a result that should
block a pipeline, since the scan results are unknown, and may be
ok on retry.

In this case, it would be useful to be able to retry or soft_fail
the buildkite step. To this end, an explicit exit code for timeouts
is added.

Co-authored-by: Amy Golshan <amy.golshan@buildkite.com>
Co-authored-by: Lucas Wilson-Richter <lucas@buildkite.com>